requirements.txt
openai < 1.0.0
langchain
python-dotenv
pypdf
chromadb
tiktoken